autogpt/requirements.txt
openai>=1.52.0
python-dotenv>=1.0.1
requests>=2.32.3
duckduckgo-search>=6.2.10
pydantic>=2.9.2
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0

